+++ src/vue/App.vue
@@ -67,6 +67,7 @@ import { vueRoutes } from '@/vue-router/routes'
 
 import config from '@/config'
 import { i18n } from '@/i18n'
+import { keyValues } from '@/key-values'
 
 export default {
   name: 'app',
@@ -123,7 +124,6 @@ export default {
 
   methods: {
     ...mapActions({
-      loadKvEntries: vuexTypes.LOAD_KV_ENTRIES,
       loadAssets: vuexTypes.LOAD_ASSETS,
       loadAccount: vuexTypes.LOAD_ACCOUNT,
       decryptSecretSeed: vuexTypes.DECRYPT_SECRET_SEED,
@@ -141,7 +141,7 @@ export default {
       const { data: networkDetails } = await api.getRaw('/')
       api.useNetworkDetails(networkDetails)
 
-      await this.loadKvEntries()
+      await keyValues.load()
 
       if (this[vuexTypes.isLoggedIn]) {
         await this.restoreSession()