+++ src/vue/forms/AssetForm/components/AdvancedStep.vue
@@ -20,7 +20,7 @@
         />
       </template>
 
-      <template v-if="kvBridgesEnabled">
+      <template v-if="$kv.bridgesEnabled > 0">
         <stellar-section
           class="app__form advanced-step__section"
           :former="former"
@@ -79,8 +79,6 @@
 <script>
 import FormMixin from '@/vue/mixins/form.mixin'
 import { AssetFormer } from '@/js/formers/AssetFormer'
-import { vuexTypes } from '@/vuex'
-import { mapGetters } from 'vuex'
 import IssuanceSection from './IssuanceSection'
 import PermissionsSection from './PermissionsSection'
 import StellarSection from './StellarSection'
@@ -116,10 +114,6 @@ export default {
       collectCallbacks: [],
       INTEGRATIONS,
     }
-  },
-
-  computed: {
-    ...mapGetters({ kvBridgesEnabled: vuexTypes.kvBridgesEnabled }),
   },
 
   methods: {