+++ src/vue/modules/assets/my-assets-explorer/index.vue
@@ -20,8 +20,6 @@
           <asset-attributes-viewer
             :asset="selectedBalance.asset"
             :balance="selectedBalance.balance"
-            :kyc-required-asset-type="kvAssetTypeKycRequired"
-            :security-asset-type="kvAssetTypeSecurity"
           />
 
           <div class="my-assets-explorer__actions-wrapper">
@@ -133,8 +131,6 @@ export default {
   computed: {
     ...mapGetters([
       vuexTypes.accountId,
-      vuexTypes.kvAssetTypeKycRequired,
-      vuexTypes.kvAssetTypeSecurity,
       vuexTypes.defaultQuoteAsset,
       vuexTypes.accountOwnedAssetsBalances,
       vuexTypes.accountBalances,