+++ src/vuex/assets.module.js
@@ -89,9 +89,9 @@ export const getters = {
       .filter(item => item.isWithdrawable),
   [vuexTypes.statsQuoteAsset]: (a, getters, b, rootGetters) =>
     rootGetters[vuexTypes.assets]
-      .filter(item => {
-        return item.isStatsQuoteAsset
-      })[0] || {},
+      .find(item => item.isStatsQuoteAsset) || {},
+  [vuexTypes.defaultQuoteAsset]: (a, getters, b, rootGetters) =>
+    rootGetters[vuexTypes.statsQuoteAsset].code,
   [vuexTypes.ownedBalancesAssets]: (a, getters, b, rootGetters) =>
     rootGetters[vuexTypes.accountBalances]
       .map(item => item.asset)